[level 1] Title: 두 개 뽑아서 더하기, Time: 2.42 ms, Memory: 80.7 MB -BaekjoonHub

+++ 프로그래머스/1/68644. 두 개 뽑아서 더하기/두 개 뽑아서 더하기.java
@@ -12,7 +12,7 @@ class Solution {
         Iterator<Integer> it = set.iterator();
         int index = 0;
         while(it.hasNext()){
-            answer[index] = (int)it.next();
+            answer[index] = it.next();
             index++;
         }
         return answer;
